[Madhav] Add: refinement
