+++ src/main/java/com/cg/gotstock/service/PortfolioService.java
@@ -5,6 +5,7 @@ import com.cg.gotstock.model.StockHolding;
 import com.cg.gotstock.model.User;
 import com.cg.gotstock.repository.StockHoldingRepository;
 import com.cg.gotstock.repository.UserRepository;
+import com.cg.gotstock.utility.PdfGenerator;
 import com.lowagie.text.DocumentException;
 import jakarta.mail.MessagingException;
 import lombok.extern.slf4j.Slf4j;
@@ -13,7 +14,6 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
