+++ src/main/java/com/cg/gotstock/service/UserService.java
@@ -156,4 +156,6 @@ public class UserService implements UserInterface {
     public boolean existsByEmail(String email) {
         return false;
     }
+
+
 }
